# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Video Generation API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/video/v1/auth/pair": {
      "post": {
        "tags": [
          "identity"
        ],
        "summary": "Pair",
        "operationId": "pairDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "default": ""
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldIssue"
            },
            "type": "array",
            "title": "Fields"
          },
          "current_row_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Row Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "FieldIssue": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldIssue"
      },
...
      "Identity": {
        "properties": {
          "actor_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Actor Id"
          },
          "tenant_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Tenant Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "workspace_name": {
            "type": "string",
            "title": "Workspace Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "editor",
              "reviewer",
              "viewer",
              "operator"
            ],
            "title": "Role"
          },
          "session_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "actor_id",
          "tenant_id",
          "display_name",
          "workspace_name",
          "role",
          "session_id"
        ],
        "title": "Identity"
      },
...
      "PairRequest": {
        "properties": {
          "pairing_code": {
            "type": "string",
            "maxLength": 256,
            "minLength": 16,
            "title": "Pairing Code"
          },
          "device_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Device Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pairing_code",
          "device_name"
        ],
        "title": "PairRequest"
      },
      "PairResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "identity": {
            "$ref": "#/components/schemas/Identity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "expires_at",
          "identity"
        ],
        "title": "PairResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}